requirements.txt
openvino-genai
openvino-tokenizers[transformers]
optimum-intel[openvino]
optimum
nncf

onnx
onnxruntime
onnxruntime-openvino
torch
transformers
huggingface_hub
tensorflow-intel
accelerate

click

vllm
